refactor: split interface in separate file
